Improve processing time of process_nifti_upload
The main improvements come from:
- Ensuring that each nifti file is only loaded once
- Working over the non-zero indices of each file (since segmentation
masks tend to be quite sparse, this gives a large improvement)
- Instead of iterating over each index (i, j, k), we group together the
indices that belong to the same group and update them all at once
- Avoid pre-converting binary masks (we can just set the instance number
for all non-zero indices when processing that file)

I have added detailed comments to explain the updated logic.

There was also a bug with the logic that decides to convert the base
image data to a uint8 array that I fixed.

For some sample data we had with 6 binary masks of dimension
(512x512x812), the processing used to take ~240 seconds and now it takes
~20 seconds.

For another sample with 2 non-binary masks (with a total of 6
instances), the processing went from ~55s down to ~14s.