Fix bug in process_nifti_upload
In the recent update, there was a bug introduced that meant group
numbers were conflicting with instance numbers. This was because we
subtracted the instance_keys from the instance_pool before populating
the instance_keys variable.

I have updated the unit test to check for this condition.

Additionally, I have added a minor optimization to the setting of
used_instances from the base_data to only call unique on non-zero
values. This saves a few seconds on large sparse segmentations.